Updated for revision
